Move from vendor/assets to app/assets for Rails 4, add css_ua.js
In Rails 4, assets like fonts and images in lib and vendor are no longer precompiled by default. Moved the skeuocard assets to app/assets so that they will be precompiled by default in Rails 4.

Also, the css browser hacks require css_ua.js to work properly. The file is included in the skeuocard project (in javascripts/vendor) but not in skeuocard-rails. Added it so that it can be required along with skeuocard.js.

Edit README to explain need for css_ua and how to include it.